BUGFIX: create new connection_set class that tracks all open connections
There were concurrency problems when dealer was waiting for an anchor
to connect and the receiver tried to send a notification that another anchor
disconnected.
This is fixed by using 2 different mutexes for thread sync and connections
data storage.

When destroying the dealer previous version removed map elements while
iterating over it causing read access violation.